--- a/resources/graph.js
+++ b/resources/graph.js
@@ -3280,7 +3280,7 @@
 	 */
 	function Picker(g){
 	
-		this.log = new Logger({'id':'Picker','logging':true,'logtime':true});
+		this.log = new Logger({'id':'Picker','logging':g.logging,'logtime':g.logtime});
 		
 		function inside(point, vs){
 			// ray-casting algorithm based on http://www.ecse.rpi.edu/Homepages/wrf/Research/Short_Notes/pnpoly.html