--- a/tests/webgl.html
+++ b/tests/webgl.html
@@ -20,7 +20,7 @@
 <body>
 	<script>
 	var range = { 'x': {'min':0,'max':10}, 'y': {'min':0,'max':10} };
-	var numperlayer = 10;
+	var numperlayer = 100;
 
 	function makeCurve(n){
 		var data = new Array(n);