… resources/timeseries.js
… tests/picker.html
… tests/webgl-debug.js
… tests/webgl-helper.js