This changes how we identify which data should be highlighted. Rather than a pixel-based lookup - calculated on each draw - we now calculate highlighted data as the cursor position changes. This speeds up the draw step. For the "basic" example I get a change from ~70ms to ~30ms and for the "microlensing" example it drops from ~330ms to ~210ms. This change was needed for the WebGL version (not yet ready) but it can be included in the 2D <canvas> version to give a performance improvement.