@astrofrog 2020-01-10
This looks good so I'll go ahead and merge - thanks!